Remove registration.cpp
It was confusing, since we have registration_impl.hpp now.
