+++ CMakeLists.txt
@@ -29,7 +29,6 @@ add_library(Library-C++
     src/affine.cpp
     src/nonrigid.cpp
     src/normalization.cpp
-    src/registration.cpp
     src/rigid.cpp
     src/utils.cpp
     ${PROJECT_BINARY_DIR}/src/version.cpp
