+++ src/utils.hpp
@@ -23,4 +23,6 @@ namespace cpd {
 
 Matrix construct_affinity_matrix(const MatrixRef source, const MatrixRef target,
                                  double beta);
+
+double default_sigma2(const MatrixRef source, const MatrixRef target);
 }
