+++ test/rigid_test.cpp
@@ -1,4 +1,5 @@
 #include "cpd/rigid.hpp"
+#include "registration_impl.hpp"
 #include "test/support.hpp"
 
 namespace cpd {
